Component-based Design 구현본
* CBD로 설계된 구현 결과에 간단하게 하나의 게임 객체 test와 그 자식 객체를 등록합니다.
* test 객체는 Transform, Renderer, MoveScript 컴포넌트로 구성되어 있음. MoveScript에서는 WASD를 이용한 방향 이동 구현
* test의 자식 객체를 만들어 Transform, Renderer, RendererEnabler 컴포넌트로 구성되어 있음. RendererEnabler는 왼쪽 마우스 버튼을 누르면 자식 객체를 보이도록, 혹은 안보이도록 제어할 수 있음
* 부모 이동시 자식 이동을 구현하기 위해 updatePos 로직을 update내에 적용함.
* Renderer 콤포넌트의 화면 그리기를 일반 콤포넌트 update 로직과 분리하여 render 로직에 넣고 이를 호출하는 internalRender 함수를 모든 컴포넌트의 update후에 수행되도록 조정함

+++ OOP2021_1/Input.cpp
@@ -20,7 +20,7 @@ bool Input::getKey(WORD virtualKeyCode) const
 {
 	if (cNumRead == 0) return false;
 
-	for (int i = 0; i < cNumRead; i++)
+	for (int i = 0; i < (int)cNumRead; i++)
 	{
 		if (irInBuf[i].EventType != KEY_EVENT) continue;
 
@@ -36,7 +36,7 @@ bool Input::getKeyUp(WORD virtualKeyCode) const
 {
 	if (cNumRead == 0) return false;
 
-	for (int i = 0; i < cNumRead; i++)
+	for (int i = 0; i < (int)cNumRead; i++)
 	{
 		if (irInBuf[i].EventType != KEY_EVENT) continue;
 
@@ -50,7 +50,7 @@ bool Input::getKeyUp(WORD virtualKeyCode) const
 
 bool Input::getMouseButtonDown(unsigned short number) 
 {
-	for (int i = 0; i < cNumRead; i++) {
+	for (int i = 0; i < (int)cNumRead; i++) {
 		if (irInBuf[i].EventType != MOUSE_EVENT) continue;
 
 		auto& mouseEvent = irInBuf[i].Event.MouseEvent;
